relocate/rename PersistentStorage to ExamplePersistentStorage
Java matter controller app test needs the example persistent storage, relocate the PersistentStorage from examples/chip-tool to src/controller, and put it nearby ExampleOperationalCredential
