+++ examples/tv-casting-app/tv-casting-common/BUILD.gn
@@ -43,6 +43,8 @@ chip_data_model("tv-casting-common") {
     "${chip_root}/examples/chip-tool/commands/common/Commands.cpp",
     "${chip_root}/examples/chip-tool/commands/common/Commands.h",
     "${chip_root}/examples/chip-tool/commands/common/CredentialIssuerCommands.h",
+    "${chip_root}/src/controller/ExamplePersistentStorage.cpp",
+    "${chip_root}/src/controller/ExamplePersistentStorage.h",
     "${chip_root}/zzz_generated/chip-tool/zap-generated/cluster/ComplexArgumentParser.cpp",
     "${chip_root}/zzz_generated/chip-tool/zap-generated/cluster/logging/DataModelLogger.cpp",
     "commands/clusters/ModelCommand.cpp",
@@ -80,6 +82,7 @@ chip_data_model("tv-casting-common") {
 
   deps = [
     "${chip_root}/src/app/tests/suites/commands/interaction_model",
+    "${chip_root}/third_party/inipp",
     "${chip_root}/third_party/jsoncpp",
   ]
